requirements.txt
pynput
google-genai